update llm_config
update llm_config
